attempted lab07 struggling hard

+++ src/routes/food.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const express = require('express');
+const user = require('../models/user/model');
 
 const FoodCollection = require('../models/index.js').Food;
 
@@ -28,22 +29,31 @@ async function getOneFood(req, res) {
 }
 
 async function createFood(req, res) {
+
   let obj = req.body;
   let newFood = await FoodCollection.create(obj);
   res.status(200).json(newFood);
 }
 
+
 async function updateFood(req, res) {
+
   const id = req.params.id;
   const obj = req.body;
   let updatedFood = await FoodCollection.update(id, obj);
   res.status(200).json(updatedFood);
 }
 
+
 async function deleteFood(req, res) {
-  let id = req.params.id;
-  let deletedFood = await FoodCollection.delete(id);
-  res.status(200).json(deletedFood);
+  if (user.role !== 'admin') {
+    res.status(403).send('Unauthorized');
+  }
+  else {
+    let id = req.params.id;
+    let deletedFood = await FoodCollection.delete(id);
+    res.status(200).json(deletedFood);
+  }
 }
 
 module.exports = router;
